Refactor: MealContainer에서 Loader와 MealCard 분리
