+++ FE/src/Loader.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Loader() {
+  return <div>Loading...</div>;
+}
+
+export default Loader;
